divOps: custom docs blocks output autodocs page
