--- packages/ui/.storybook/preview.ts
@@ -1,16 +1,0 @@
-import type { Preview } from "@storybook/react";
-
-import "../dist/output.css";
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
